+++ src/test/comparison/comparison.h
@@ -1,11 +1,11 @@
 #include <stdbool.h>
 
-bool eq(int x, int y);
-bool neq(int x, int y);
-bool lt(int x, int y);
-bool lte(int x, int y);
-bool gt(int x, int y);
-bool gte(int x, int y);
+bool eq(long x, long y);
+bool neq(long x, long y);
+bool lt(long x, long y);
+bool lte(long x, long y);
+bool gt(long x, long y);
+bool gte(long x, long y);
 
 bool and(bool a, bool b);
 bool or(bool a, bool b);
