+++ src/test/math/math.h
@@ -2,28 +2,28 @@
   Die nachfolgenden Funktionen werden später dazu gelinkt
   Der Ursprung ist die tests.k Datei, dort sind diese Funktionen implementiert
 */
-int add(int x, int y);
-int sub(int x, int y); 
-int mul(int x, int y);
-int modulo(int x, int y);
-int neg(int x);
-int id(int x);
-int selfMinus(int x);
-int t1();
-int t2();
-int t3();
-int t4();
-int t5();
-int t6();
-int t7();
-int t8();
+long add(long x, long y);
+long sub(long x, long y); 
+long mul(long x, long y);
+long modulo(long x, long y);
+long neg(long x);
+long id(long x);
+long selfMinus(long x);
+long t1();
+long t2();
+long t3();
+long t4();
+long t5();
+long t6();
+long t7();
+long t8();
 double fadd(double x, double y);
 double fsub(double x, double y);
 double fmul(double x, double y);
 double fneg(double x);
 double fid(double x);
 double fselfMinus(double x);
-double mixadd(double x, int y);
-double mixsub(double x, int y);
-double mixmul(double x, int y);
-double mixdiv(double x, int y);
+double mixadd(double x, long y);
+double mixsub(double x, long y);
+double mixmul(double x, long y);
+double mixdiv(double x, long y);
