+++ src/test/recursive/recursive.h
@@ -1,1 +1,1 @@
-int fac(int x);
+long fac(long x);
